финал
Почему-то get_header() иногда  не срабатывает
